fix/ rename defmodule Forecast -> ForecastAggregator and aggretate moon status and tide name handling
